+++ tennis/tennis.rb
@@ -1,76 +1,5 @@
 require "pry"
 
-module Tennis
-  class Game1
-  end
-
-  class Player
-    def initialize(name:, points: 0)
-      @name = name
-      @points = points
-    end
-
-    def won
-      @points += 1
-    end
-
-    attr_reader :name, :points
-  end
-end
-
-class TennisGame1
-  def initialize(player1_name, player2_name)
-    @players = []
-    @players << @player_1 = Tennis::Player.new(name: player1_name)
-    @players << @player_2 = Tennis::Player.new(name: player2_name)
-  end
-
-  def won_point(player_name)
-    player = find_player_by_name(player_name)
-    player.won
-  end
-
-  def score
-    return draw_result if @player_1.points == @player_2.points
-    return ongoing_result unless (@player_1.points >= 4) || (@player_2.points >= 4)
-    matching_point_result
-  end
-
-  private
-
-  def matching_point_result
-    score_diff = @player_1.points - @player_2.points
-    leading_player = score_diff.positive? ? @player_1 : @player_2
-    condition = score_diff.abs == 1 ? "Advantage " : "Win for "
-    condition + leading_player.name
-  end
-
-  def ongoing_result
-    "#{score_label(@player_1.points)}-#{score_label(@player_2.points)}"
-  end
-
-  def score_label(score)
-    {
-      0 => "Love",
-      1 => "Fifteen",
-      2 => "Thirty",
-      3 => "Forty"
-    }[score]
-  end
-
-  def draw_result
-    {
-      0 => "Love-All",
-      1 => "Fifteen-All",
-      2 => "Thirty-All"
-    }.fetch(@player_1.points, "Deuce")
-  end
-
-  def find_player_by_name(name)
-    @players.find { |p| p.name == name }
-  end
-end
-
 class TennisGame2
   def initialize(player1Name, player2Name)
     @player1Name = player1Name
